Move repositories into unified namespace
