+++ site/app/config/services.yml
@@ -9,6 +9,18 @@ services:
         autoconfigure: true
         public: false
 
+    # makes classes in src/Librecores/ProjectRepoBundle available to be used as services
+    # this creates a service per class whose id is the fully-qualified class name
+    App\:
+        resource: '../../src/App/*'
+        exclude: '../../src/App/{Entity, DataFixtures}'
+
+    # controllers are imported separately to make sure services can be injected
+    # as action arguments even if you don't extend any base controller class
+#    App\Controller\:
+#        resource: '../../src/App/Controller'
+#        tags: ['controller.service_arguments']
+
     # Twig date extension - http://twig-extensions.readthedocs.io/en/latest/date.html
     Twig\Extensions\DateExtension:
         autowire: false
