+++ site/src/App/RepoCrawler/GitRepoCrawler.php
@@ -15,8 +15,8 @@ use App\Entity\GitSourceRepo;
 use App\Entity\LanguageStat;
 use App\Entity\ProjectRelease;
 use App\Entity\SourceRepo;
-use Librecores\ProjectRepoBundle\Repository\CommitRepository;
-use Librecores\ProjectRepoBundle\Repository\ContributorRepository;
+use App\Repository\CommitRepository;
+use App\Repository\ContributorRepository;
 use Psr\Log\LoggerInterface;
 use RuntimeException;
 use Symfony\Component\Filesystem\Filesystem;
