+++ site/src/App/Repository/CommitRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Librecores\ProjectRepoBundle\Repository;
+namespace App\Repository;
 
 use App\Entity\Commit;
 use App\Entity\Contributor;
