+++ site/src/App/Repository/OrganizationRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Librecores\ProjectRepoBundle\Repository;
+namespace App\Repository;
 
 use App\Entity\Organization;
 use App\Entity\User;
