+++ site/src/Librecores/ProjectRepoBundle/Controller/ProjectController.php
@@ -10,6 +10,8 @@ use App\Entity\Project;
 use App\Entity\ProjectClassification;
 use App\Entity\User;
 use App\RepoCrawler\GithubRepoCrawler;
+use App\Repository\OrganizationRepository;
+use App\Repository\ProjectRepository;
 use App\Util\Dates;
 use DateInterval;
 use DateTimeImmutable;
@@ -18,8 +20,6 @@ use Exception;
 use FOS\UserBundle\Model\UserManagerInterface;
 use Librecores\ProjectRepoBundle\Doctrine\ProjectMetricsProvider;
 use Librecores\ProjectRepoBundle\Form\Type\ProjectType;
-use Librecores\ProjectRepoBundle\Repository\OrganizationRepository;
-use Librecores\ProjectRepoBundle\Repository\ProjectRepository;
 use Librecores\ProjectRepoBundle\Service\GitHub\AuthenticationRequiredException;
 use Librecores\ProjectRepoBundle\Service\GitHub\GitHubApiService;
 use Librecores\ProjectRepoBundle\Service\QueueDispatcherService;
