+++ site/tests/RepoCrawler/GitRepoCrawlerTest.php
@@ -12,8 +12,8 @@ use App\Util\ProcessCreator;
 use DateTime;
 use Doctrine\Common\Persistence\ObjectManager;
 use Librecores\ProjectRepoBundle\Doctrine\ProjectMetricsProvider;
-use Librecores\ProjectRepoBundle\Repository\CommitRepository;
-use Librecores\ProjectRepoBundle\Repository\ContributorRepository;
+use App\Repository\CommitRepository;
+use App\Repository\ContributorRepository;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Process\Process;
