+++ site/tests/Repository/CommitRepositoryTest.php
@@ -1,33 +1,31 @@
 <?php
 
-namespace Librecores\ProjectRepoBundle\Repository;
+namespace App\Repository;
 
 
-use App\Entity\Contributor;
+use App\Entity\Commit;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Mapping\ClassMetadata;
 use PHPUnit\Framework\TestCase;
 use Symfony\Bridge\Doctrine\RegistryInterface;
 
-class ContributorRepositoryTest extends TestCase
+class CommitRepositoryTest extends TestCase
 {
-    public function testContributorRepositoryIsMappedToContributorEntity()
+    public function testCommitRepositoryIsMappedToCommitEntity()
     {
         $mockEntityManager = $this->createMock(EntityManagerInterface::class);
         $mockEntityManager->expects($this->once())
             ->method('getClassMetadata')
-            ->with(Contributor::class)
-            ->willReturn(new ClassMetadata(Contributor::class));
+            ->with(Commit::class)
+            ->willReturn(new ClassMetadata(Commit::class));
 
         $mockRegistry = $this->createMock(RegistryInterface::class);
         $mockRegistry->expects($this->once())
             ->method('getManagerForClass')
-            ->with(Contributor::class)
+            ->with(Commit::class)
             ->willReturn($mockEntityManager);
 
         /** @var RegistryInterface $mockRegistry */
-        $repository = new ContributorRepository($mockRegistry);
-
-        $this->assertEquals(Contributor::class, $repository->getClassName());
+        new CommitRepository($mockRegistry);
     }
 }
