+++ site/tests/Repository/ProjectRepositoryTest.php
@@ -1,31 +1,31 @@
 <?php
 
-namespace Librecores\ProjectRepoBundle\Repository;
+namespace App\Repository;
 
 
-use App\Entity\User;
+use App\Entity\Project;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Mapping\ClassMetadata;
 use PHPUnit\Framework\TestCase;
 use Symfony\Bridge\Doctrine\RegistryInterface;
 
-class UserRepositoryTest extends TestCase
+class ProjectRepositoryTest extends TestCase
 {
-    public function testUserRepositoryIsMappedToUserEntity()
+    public function testProjectRepositoryIsMappedToProjectEntity()
     {
         $mockEntityManager = $this->createMock(EntityManagerInterface::class);
         $mockEntityManager->expects($this->once())
             ->method('getClassMetadata')
-            ->with(User::class)
-            ->willReturn(new ClassMetadata(User::class));
+            ->with(Project::class)
+            ->willReturn(new ClassMetadata(Project::class));
 
         $mockRegistry = $this->createMock(RegistryInterface::class);
         $mockRegistry->expects($this->once())
             ->method('getManagerForClass')
-            ->with(User::class)
+            ->with(Project::class)
             ->willReturn($mockEntityManager);
 
         /** @var RegistryInterface $mockRegistry */
-        new UserRepository($mockRegistry);
+        new ProjectRepository($mockRegistry);
     }
 }
